requirements.txt
django==3.2
djangorestframework==3.12.4
flake8==3.8.4
pylint-django==2.0.15
pylint==2.5.2
iniconfig==1.0.1
mccabe==0.6.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pytest==6.2.3
pytest-cov==2.9.0
pytest-django==3.9.0
pyjwt==1.7.1
requests==2.23.0
psycopg2-binary==2.8.5
factory-boy==2.12.0